backend/requirements.txt
# Core
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Gemini AI
google-generativeai>=0.3.0

# NLP & Text Processing
spacy>=3.5.0
nltk>=3.8.1
scikit-learn>=1.2.2
numpy>=1.24.3
python-dateutil>=2.8.2

# Document Processing
pdfplumber>=0.9.0
python-docx>=0.8.11
python-magic>=0.4.27

# Voice & Audio
edge-tts>=6.1.0
gTTS>=2.3.2
SpeechRecognition>=3.10.0
pydub>=0.25.1
pyaudio>=0.2.13
pyttsx3>=2.90

# Web & Async
aiohttp>=3.8.4
requests>=2.28.2

# Database (if needed)
sqlalchemy>=2.0.0
databases[postgresql]>=0.7.0
alembic>=1.10.3
psycopg2-binary>=2.9.5

# Testing & Development
pytest>=7.3.1
pytest-cov>=4.0.0
httpx>=0.24.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0